rename compactFractionSerialization to compactSerialization, apply to \sqrt
The single-digit compact form is not specific to fractions: SurdAtom
had its own copy of the rule, so `\sqrt{2}` serialized as `\sqrt2`
regardless of the option. Gate both on a single setting.

With compactSerialization off, `\sqrt12` now serializes as `\sqrt{1}2`,
making the actual parse (root of 1, followed by 2) visible.
