+++ src/api.md
@@ -805,19 +805,18 @@ manually.
 
 <MemberCard>
 
-##### MathfieldElement.compactFractionSerialization
+##### MathfieldElement.compactSerialization
 
 ```ts
-get static compactFractionSerialization(): boolean
-set static compactFractionSerialization(value: boolean): void
+get static compactSerialization(): boolean
+set static compactSerialization(value: boolean): void
 ```
 
-When `true` (the default), fractions whose numerator and denominator
-are both single digits are serialized in a compact form, e.g.
-`\frac12`.
+When `true` (the default), single-digit arguments are serialized in a
+compact form, without braces: `\frac12`, `\sqrt2`.
 
-When `false`, fractions are always serialized with explicit braces,
-e.g. `\frac{1}{2}`.
+When `false`, arguments are always serialized with explicit braces:
+`\frac{1}{2}`, `\sqrt{2}`.
 
 **Default**: `true`
